Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/character_solver/skinned_locator_triangle_error_function.cpp
@@ -619,7 +619,7 @@ double SkinnedLocatorTriangleErrorFunctionT<T>::getJacobian(
           jacobian.block(3 * i, 0, 3, modelParameters.size()),
           residual.middleRows(3 * i, 3));
     }
-    usedRows = 3 * constraints_.size();
+    usedRows = static_cast<int>(3 * constraints_.size());
   } else {
     MT_CHECK(constraintType_ == VertexConstraintType::Plane);
     for (size_t i = 0; i < constraints_.size(); ++i) {
@@ -639,7 +639,7 @@ double SkinnedLocatorTriangleErrorFunctionT<T>::getJacobian(
           residualValue);
       residual(i) = residualValue;
     }
-    usedRows = constraints_.size();
+    usedRows = static_cast<int>(constraints_.size());
   }
 
   return error;
